# storage/api-docs/api-spec.yaml
openapi: 3.0.0
info:
  title: 'MegaSorpresa API'
  description: 'API Backend para el e-commerce de juguetes MegaSorpresa. Sirve a clientes Web (SPA), Android e iOS.'
  contact:
    email: api@megasorpresa.com
  version: 1.0.0
servers:
  -
    url: /
    description: 'API Server'
paths:
...
  /api/cart/add:
    post:
      tags:
        - Carrito
      summary: 'Agregar producto al carrito'
      description: 'Agrega un producto al carrito. Si el producto ya existe, incrementa la cantidad.'
      operationId: e940193f73a57895e17830b67c48f89c
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - product_id
                - quantity
              properties:
                product_id:
                  type: integer
                  example: 1
                quantity:
                  type: integer
                  minimum: 1
                  example: 2
              type: object
      responses:
        '200':
          description: 'Producto agregado exitosamente'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Product added to cart successfully' }
                  data: { $ref: '#/components/schemas/Cart' }
                type: object
        '422':
          description: 'Error de validación o stock insuficiente'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Failed to add product to cart' }
                  error: { type: string, example: 'Insufficient stock for product' }
                type: object
  /api/cart/update-quantity:
    patch:
      tags:
        - Carrito
      summary: 'Actualizar cantidad de un item del carrito'
      description: 'Modifica la cantidad de un producto en el carrito.'
      operationId: 76323813ed9d18249c39ac7e76d729df
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - product_id
                - quantity
              properties:
                product_id:
                  type: integer
                  example: 1
                quantity:
                  type: integer
                  minimum: 1
                  example: 3
              type: object
      responses:
        '200':
          description: 'Cantidad actualizada exitosamente'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Cart item quantity updated successfully' }
                  data: { $ref: '#/components/schemas/Cart' }
                type: object
        '404':
          description: 'Producto no encontrado en el carrito'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Product not found in cart' }
                type: object
        '422':
          description: 'Error de validación o stock insuficiente'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Failed to update cart item quantity' }
                  error: { type: string, example: 'Insufficient stock' }
                type: object
  '/api/cart/remove/{productId}':
    delete:
      tags:
        - Carrito
      summary: 'Eliminar producto del carrito'
      description: 'Remueve un producto del carrito completamente.'
      operationId: 44a0e59980c9b61b8f11b0cec3589a34
      parameters:
        -
          name: productId
          in: path
          description: 'ID del producto a eliminar'
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 'Producto eliminado exitosamente'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Product removed from cart successfully' }
                  data: { $ref: '#/components/schemas/Cart' }
                type: object
  /api/cart/details:
    post:
      tags:
        - Carrito
      summary: 'Actualizar información de envío del carrito'
      description: 'Actualiza el código postal, ciudad y fecha programada de entrega del carrito.'
      operationId: 3a1efa573bd57700f027ae6c9dd1d7a3
      requestBody:
        required: false
        content:
          application/json:
            schema:
              properties:
                shipping_zip_code:
                  type: string
                  maxLength: 10
                  example: '12345'
                shipping_city_id:
                  type: integer
                  example: 1
                scheduled_delivery_date:
                  description: 'Fecha programada de entrega (debe ser posterior a hoy)'
                  type: string
                  format: date
                  example: '2026-05-15'
              type: object
      responses:
        '200':
          description: 'Detalles actualizados exitosamente'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Cart details updated successfully' }
                  data: { $ref: '#/components/schemas/Cart' }
                type: object
        '422':
          description: 'Error de validación'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Failed to update cart details' }
                  error: { type: string, example: 'Delivery date must be in the future' }
                type: object
...
  '/api/catalog/categories/{id}':
    get:
      tags:
        - Catálogo
      summary: 'Obtener detalle de una categoría con sus productos'
      description: 'Obtiene la información de una categoría específica junto con sus productos asociados. Los productos se pueden filtrar por ciudad y están paginados.'
      operationId: de30548fbe81774b79d5d5615e90b7dd
      parameters:
        -
          name: id
          in: path
          description: 'ID de la categoría'
          required: true
          schema:
            type: integer
            example: 1
        -
          name: city_id
          in: query
          description: 'ID de la ciudad para filtrar productos por disponibilidad'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: per_page
          in: query
          description: 'Número de productos por página'
          required: false
          schema:
            type: integer
            default: 15
            example: 15
      responses:
        '200':
          description: 'Detalle de la categoría con productos obtenido exitosamente'
          content:
            application/json:
              schema:
                properties:
                  category: { $ref: '#/components/schemas/Category' }
                  products: { type: array, items: { $ref: '#/components/schemas/Product' } }
                type: object
        '404':
          description: 'Categoría no encontrada'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
...
components:
  schemas:
    CartItem:
      title: CartItem
      description: 'Item dentro del carrito de compras'
      properties:
        id:
          type: integer
          example: 1
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: integer
          example: 2
        price_at_addition:
          type: number
          format: float
          example: 29.99
        subtotal:
          type: number
          format: float
          example: 59.98
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    Cart:
      title: Cart
      description: 'Carrito de compras'
      properties:
        id:
          type: integer
          example: 1
        session_id:
          type: string
          example: abc123
        shipping_zip_code:
          type: string
          example: '12345'
          nullable: true
        shipping_city:
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: 'Ciudad de México'
          type: object
          nullable: true
        scheduled_delivery_date:
          type: string
          format: date
          example: '2026-05-15'
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: number
          format: float
          example: 59.98
        total_items:
          type: integer
          example: 2
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    Product:
      title: Product
      description: 'Modelo de producto'
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Muñeca Barbie'
        slug:
          type: string
          example: muneca-barbie
        sku:
          type: string
          example: TOY-001
        base_price:
          type: number
          format: float
          example: 29.99
        description:
          type: string
          example: 'Muñeca Barbie original con accesorios'
        stock_quantity:
          type: integer
          example: 50
        is_active:
          type: boolean
          example: true
        primary_image:
          type: string
          example: 'https://example.com/images/product.jpg'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      type: object
    Category:
      title: Category
      description: 'Modelo de categoría'
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Muñecas
        slug:
          type: string
          example: munecas
        description:
          type: string
          example: 'Muñecas y accesorios'
        is_active:
          type: boolean
          example: true
      type: object
...
    NotFoundError:
      title: 'Not Found Error'
      description: 'Recurso no encontrado (404)'
      properties:
        message:
          type: string
          example: 'Resource not found.'
      type: object
...
  securitySchemes:
    sanctum:
      type: http
      description: 'Token de autenticación de Laravel Sanctum para clientes móviles'
      bearerFormat: Token
      scheme: bearer
...